# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Days Off App",
    "description": "Thanks to this application , the leave management system empowers your staff on . question such as How many days of paid vacation do i have left?  Your staff can check balances ,manage and track their time off.Managers can run reports,analyze absence trends,and find solutions to staffing issues in real time."
  },
  "servers": [
    {
      "url": "http://localhost:8000/"
    }
  ],
...
  "paths": {
...
    "/users/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Display user",
        "description": "Display user",
        "operationId": "DisplayUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request "
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Update user",
        "description": "Update user",
        "operationId": "UpdateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "requestBody": {
          "description": "Created user object",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "firstName"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "lastName"
                  },
                  "phone": {
                    "type": "string",
                    "example": "phone"
                  },
                  "avatar": {
                    "type": "string",
                    "example": "avatar"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request "
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Delete user",
        "description": "Delete user",
        "operationId": "DeleteUsers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The name that needs to be deleted",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "firstName lastName is succussffully deleted"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "required": [
          "firstName",
          "lastName",
          "email",
          "role",
          "building",
          "phone"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "exemple"
          },
          "lastName": {
            "type": "string",
            "example": "exemple"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "exemple@email.com"
          },
          "phone": {
            "type": "string",
            "example": "2345"
          },
          "role": {
            "type": "string",
            "example": "Software Engineer",
            "enum": [
              "Director",
              "Administration Director",
              "Administration Assistant",
              "Team manager",
              "Software Engineer",
              "Super Admin"
            ]
          },
          "building": {
            "type": "string",
            "example": "Front-End",
            "enum": [
              "Front-End",
              "Back-End",
              "Full-Stack"
            ]
          },
          "avatar": {
            "type": "string",
            "example": "12345"
          },
          "isActive": {
            "type": "boolean",
            "example": "true"
          },
          "soldDays": {
            "type": "Number",
            "default": "0"
          },
          "allDaysoff": {
            "type": "Number",
            "default": "0"
          },
          "daysOffSick": {
            "type": "Number",
            "default": "0"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  }
}